lr4/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib
python-multipart
pymongo